[PDF-113] Refactor WatermarkTab into smaller subcomponents
Split the ~650-line WatermarkTab.tsx into focused subcomponents:
- WatermarkTextConfig: text input, font size, font color, opacity fields
- WatermarkLayoutConfig: rotation, position, font family selects
- WatermarkPageRange: all/specific radio, page range input with validation

State management remains in the parent WatermarkTab (~370 lines).
Subcomponents are controlled components receiving values and callbacks.
The validatePageRange function moved to WatermarkPageRange module.
positionOptions and fontFamilyOptions are constants in their respective files.
getFontStyle helper lives in WatermarkLayoutConfig.

Closes #113
