+++ pdf_wizard/frontend/src/components/watermark/index.ts
@@ -1,0 +1,3 @@
+export { WatermarkTextConfig } from './WatermarkTextConfig';
+export { WatermarkLayoutConfig } from './WatermarkLayoutConfig';
+export { WatermarkPageRange, validatePageRange } from './WatermarkPageRange';
